QueryParamName.js moved to common catalog
